refactor: reorganize Npgsql plugin structure
